+++ src/Pages/Notification/Notification.jsx
@@ -1,84 +1,10 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import { Bell, Check } from "lucide-react";
-import axiosSecure from "../../utils/axiosSecure.js";
+import { useNotifications } from "../../utils/hooks/useNotifications.js";
+import CardSkeleton from "../../Components/CardSkeleton/CardSkeleton.jsx";
 
 const Notifications = () => {
-    const [notifications, setNotifications] = useState([]);
-    const [loading, setLoading] = useState(true);
-
-    const timeAgo = (dateString) => {
-        const now = new Date();
-        const past = new Date(dateString);
-        const diffInSeconds = Math.floor((now - past) / 1000);
-        if (diffInSeconds < 60) return "Just now";
-        const diffInMinutes = Math.floor(diffInSeconds / 60);
-        if (diffInMinutes < 60) return `${diffInMinutes} min${diffInMinutes > 1 ? "s" : ""} ago`;
-        const diffInHours = Math.floor(diffInMinutes / 60);
-        if (diffInHours < 24) return `${diffInHours} hr${diffInHours > 1 ? "s" : ""} ago`;
-        const diffInDays = Math.floor(diffInHours / 24);
-        if (diffInDays < 7) return `${diffInDays} day${diffInDays > 1 ? "s" : ""} ago`;
-        const diffInWeeks = Math.floor(diffInDays / 7);
-        if (diffInWeeks < 4) return `${diffInWeeks} week${diffInWeeks > 1 ? "s" : ""} ago`;
-        const diffInMonths = Math.floor(diffInDays / 30);
-        if (diffInMonths < 12) return `${diffInMonths} month${diffInMonths > 1 ? "s" : ""} ago`;
-        const diffInYears = Math.floor(diffInDays / 365);
-        return `${diffInYears} year${diffInYears > 1 ? "s" : ""} ago`;
-    };
-
-    const isToday = (dateString) => {
-        const created = new Date(dateString);
-        const today = new Date();
-        return (
-            created.getFullYear() === today.getFullYear() &&
-            created.getMonth() === today.getMonth() &&
-            created.getDate() === today.getDate()
-        );
-    };
-
-    useEffect(() => {
-        const fetchNotifications = async () => {
-            try {
-                const res = await axiosSecure.get("/notifications");
-                const formatted = (res.data.notifications || []).map((n, index) => ({
-                    id: index + 1,
-                    _id: n._id,
-                    title: n.type,
-                    message: n.message,
-                    time: timeAgo(n.createdAt),
-                    read: n.isRead,
-                    today: isToday(n.createdAt)
-                }));
-                setNotifications(formatted);
-            } catch (error) {
-                console.log(error);
-            } finally {
-                setLoading(false);
-            }
-        };
-        fetchNotifications();
-    }, []);
-
-    const unreadCount = notifications.filter(n => !n.read).length;
-    const todayNotifs = notifications.filter(n => n.today);
-    const historyNotifs = notifications.filter(n => !n.today);
-
-    const markAllRead = async () => {
-        try {
-            await axiosSecure.patch("/notifications/all");
-            setNotifications(prev => prev.map(n => ({ ...n, read: true })));
-        } catch (error) {
-            console.log(error);
-        }
-    };
-
-    const markRead = async (id, _id) => {
-        try {
-            await axiosSecure.patch(`/notifications/${_id}`);
-            setNotifications(prev => prev.map(n => n.id === id ? { ...n, read: true } : n));
-        } catch (error) {
-            console.log(error);
-        }
-    };
+    const { notifications, unreadCount, todayNotifs, historyNotifs, markAllRead, markRead, loading } = useNotifications();
 
     const NotificationRow = ({ n }) => (
         <div
@@ -125,25 +51,10 @@ const Notifications = () => {
                 </div>
 
                 {/* Loading skeleton */}
-                {loading && (
-                    <div className="bg-white rounded-2xl border border-gray-200 overflow-hidden">
-                        <div className="px-4 py-3 border-b border-gray-100 bg-gray-50">
-                            <div className="h-3 w-12 bg-gray-200 rounded animate-pulse" />
-                        </div>
-                        <div className="p-3 space-y-2">
-                            {[...Array(4)].map((_, i) => (
-                                <div key={i} className="flex items-start gap-4 p-4 rounded-xl border border-gray-100">
-                                    <div className="mt-1.5 w-2 h-2 rounded-full bg-gray-200 shrink-0" />
-                                    <div className="flex-1 space-y-2">
-                                        <div className="h-3 bg-gray-200 rounded animate-pulse w-1/3" />
-                                        <div className="h-3 bg-gray-100 rounded animate-pulse w-2/3" />
-                                    </div>
-                                    <div className="h-3 w-12 bg-gray-100 rounded animate-pulse" />
-                                </div>
-                            ))}
-                        </div>
-                    </div>
-                )}
+
+                {
+                    loading && <CardSkeleton variant="notification"></CardSkeleton>
+                }
 
                 {/* Today */}
                 {!loading && todayNotifs.length > 0 && (
